Corrected force_task_save checkpointing

+++ glompo/core/manager.py
@@ -20,7 +20,6 @@ from typing import Any, Callable, Dict, List, Optional, Sequence, Set, Tuple
 
 import numpy as np
 import yaml
-from core.scope import GloMPOScope
 
 try:
     import dill
@@ -508,7 +507,7 @@ class GloMPOManager:
         # Setup Task
         try:
             self.task = None
-            if 'task' in os.listdir('/tmp/glompo_chkpt') and not self.checkpoint_options.force_task_save:
+            if 'task' in os.listdir('/tmp/glompo_chkpt'):
                 with open('/tmp/glompo_chkpt/task', 'rb') as file:
                     try:
                         self.task = dill.load(file)
@@ -984,22 +983,28 @@ class GloMPOManager:
                 self.logger.debug("Manager successfully pickled")
 
                 # Save task
-                try:
-                    with open('task', 'wb') as file:
-                        dill.dump(self.task, file)
+                task_persisted = False
+
+                if not self.checkpoint_options.force_task_save:
+                    try:
+                        with open('task', 'wb') as file:
+                            dill.dump(self.task, file)
                         self.logger.info("Task successfully pickled")
-                except PickleError as pckl_err:
-                    self.logger.info(f"Pickle task failed: {pckl_err}. Attempting task.checkpoint_save()")
-                    os.remove('task')
+                        task_persisted = True
+                    except PickleError as pckl_err:
+                        self.logger.info(f"Pickle task failed: {pckl_err}. Attempting task.checkpoint_save()")
+                        os.remove('task')
+
+                if not task_persisted:
                     try:
                         # noinspection PyUnresolvedReferences
                         self.task.checkpoint_save('task')
                         self.logger.info("Task successfully saved")
                     except AttributeError:
-                        self.logger.debug(f"task.checkpoint_save not found.")
+                        self.logger.info(f"task.checkpoint_save not found.")
                         self.logger.warning("Checkpointing without task.")
                     except Exception as e:
-                        self.logger.debug(f"task.checkpoint_save failed", exc_info=e)
+                        self.logger.warning("Task saving failed", exc_info=e)
                         self.logger.warning("Checkpointing without task.")
 
             # Save scope
